refactor and add to src-2
